Refactor detection phase: Consolidate type definitions and exception handling in `src/types.py` and `src/detection` modules. Migrate hand-related classes to `types.py` for better organization and maintainability. Update logging practices to replace print statements with logger calls, enhancing error tracking. Address mypy strict compliance and improve overall code structure for better reliability and performance.
